// front-end/src/api/models/agendamentoRespostaDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Agendador de Barbearia
 * Contrato OpenAPI da API REST do Agendador de Barbearia.
 * OpenAPI spec version: 1.0
 */
import type { AgendamentoRespostaDtoStatus } from './agendamentoRespostaDtoStatus';
import type { BarbeiroAgendamentoRespostaDto } from './barbeiroAgendamentoRespostaDto';
import type { FilialPublicaDto } from './filialPublicaDto';
import type { ItemAgendamentoRespostaDto } from './itemAgendamentoRespostaDto';

export interface AgendamentoRespostaDto {
  id: number;
  inicioPrevisto: string;
  fimPrevisto: string;
  status: AgendamentoRespostaDtoStatus;
  /** @nullable */
  observacaoCliente?: string | null;
  /** @nullable */
  motivoCancelamento?: string | null;
  filial: FilialPublicaDto;
  barbeiro: BarbeiroAgendamentoRespostaDto;
  servicos: ItemAgendamentoRespostaDto[];
}